Revert "Feature/#23 마이페이지api"
