+++ build.gradle.kts
@@ -43,9 +43,6 @@ dependencies {
 
     runtimeOnly ("io.jsonwebtoken:jjwt-impl:0.11.2")
     runtimeOnly("org.postgresql:postgresql")
-
-    implementation(platform("com.amazonaws:aws-java-sdk-bom:1.11.228"))
-    implementation ("com.amazonaws:aws-java-sdk-s3")
     testImplementation("org.springframework.boot:spring-boot-starter-test") {
         exclude(group = "org.junit.vintage", module = "junit-vintage-engine")
         exclude(group = "org.mockito")
