--- src/main/kotlin/com/yapp/web2/domain/account/controller/AccountController.kt
@@ -1,59 +1,0 @@
-package com.yapp.web2.domain.account.controller
-
-import com.yapp.web2.domain.account.entity.Account
-import com.yapp.web2.domain.account.service.AccountService
-import com.yapp.web2.security.jwt.TokenDto
-import com.yapp.web2.util.Message
-import io.swagger.annotations.ApiOperation
-import io.swagger.annotations.ApiParam
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.*
-import org.springframework.web.multipart.MultipartFile
-import javax.servlet.http.HttpServletRequest
-import javax.validation.Valid
-
-@RestController
-@RequestMapping("/api/v1/user")
-class AccountController(
-    private val accountService: AccountService
-) {
-    @PostMapping("/oauth2Login")
-    fun oauth2Login(
-        @RequestBody @ApiParam(value = "회원 정보", required = true) request: Account.AccountLoginRequest
-    ): ResponseEntity<TokenDto> {
-        val tokenDto = accountService.oauth2LoginUser(request)
-        return ResponseEntity.status(HttpStatus.OK).body(tokenDto)
-    }
-
-    @ApiOperation(value = "토큰 재발급")
-    @GetMapping("/reIssuanceAccessToken")
-    fun reIssuanceAccessToken(request: HttpServletRequest): ResponseEntity<TokenDto> {
-        val accessToken = request.getHeader("AccessToken")
-        val refreshToken = request.getHeader("RefreshToken")
-        val tokenDto = accountService.reIssuedAccessToken(accessToken, refreshToken)
-        return ResponseEntity.status(HttpStatus.OK).body(tokenDto)
-    }
-
-    @PostMapping("/changeProfileImage")
-    fun changeProfileImage(request: HttpServletRequest, @RequestParam("images") multipartFile: MultipartFile): ResponseEntity<String> {
-        val token = request.getHeader("AccessToken")
-        val image = accountService.changeProfileImage(token, multipartFile)
-        return ResponseEntity.status(HttpStatus.OK).body(image)
-    }
-
-    @GetMapping("deleteProfileImage")
-    fun deleteProfileImage(request: HttpServletRequest): ResponseEntity<String> {
-        val token = request.getHeader("AccessToken")
-        accountService.deleteProfileImage(token)
-        return ResponseEntity.status(HttpStatus.OK).body(Message.SUCCESS)
-    }
-
-    @PostMapping("/nickNameCheck")
-    fun nickNameCheck(@RequestParam("nickName") @Valid nickName: Account.nextNickName): ResponseEntity<String> {
-        accountService.checkNickNameDuplication(nickName)
-        return ResponseEntity.status(HttpStatus.OK).body(Message.SUCCESS)
-    }
-
-
-}
