+++ src/main/kotlin/com/yapp/web2/domain/account/entity/Account.kt
@@ -5,7 +5,6 @@ import com.yapp.web2.domain.folder.entity.AccountFolder
 import com.yapp.web2.domain.notification.entity.Notification
 import javax.persistence.*
 import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.Size
 
 @Entity
 class Account(
@@ -16,8 +15,6 @@ class Account(
         fun requestToAccount(dto: AccountLoginRequest): Account {
             return Account(dto.email, dto.imageUrl, dto.name, dto.socialType)
         }
-
-        const val BASIC_IMAGE_URL: String = "https://yapp-bucket-test.s3.ap-northeast-2.amazonaws.com/basicImage.png"
     }
 
     constructor(email: String, image: String, nickname: String, socialType: String) : this(email) {
@@ -33,7 +30,7 @@ class Account(
     var socialType: String = "none"
 
     // TODO: 디폴트 사진 url 추가하기
-    var image: String = BASIC_IMAGE_URL
+    var image: String = ""
 
     var backgroundColor: String = "black"
     var remindToggle: Boolean = true
@@ -53,10 +50,5 @@ class Account(
         val imageUrl: String,
         @field: NotEmpty(message = "소셜타입을 입력해주세요")
         val socialType: String
-    )
-
-    class nextNickName(
-        @field: Size(max = 20)
-        val nickName: String
     )
 }
