+++ src/main/kotlin/com/yapp/web2/domain/account/repository/UserRepository.kt
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AccountRepository : JpaRepository<Account, Long> {
+interface UserRepository : JpaRepository<Account, Long> {
     fun findByEmail(email: String): Account?
-    fun findByNickname(nickName: String): Account?
 }
