--- src/test/kotlin/com/yapp/web2/config/S3UploaderTest.kt
@@ -1,12 +1,0 @@
-package com.yapp.web2.config
-
-import org.junit.jupiter.api.Test
-
-import org.junit.jupiter.api.Assertions.*
-
-internal class S3UploaderTest {
-
-    @Test
-    fun upload() {
-    }
-}
